requirements.txt
# Pip
pip~=23.0.1

# Dataframe
numpy~=1.24.2
pandas~=1.4.4
openpyxl~=3.1.2

# Streamlit
streamlit~=1.19.0

# Visualization
plotly_express~=0.4.1
kaleido~=0.2.1
umap-learn~=0.5.3

# Machine Learning
torch~=1.13.0
sentence-transformers~=2.2.2
pytorch-lightning~=1.9.4
torchmetrics~=0.11.3
transformers~=4.26.1

# Logging ML
tensorboard~=2.12.0
